+++ StreamDeckVSC/Configuration.cs
@@ -1,0 +1,9 @@
+﻿namespace StreamDeckVSC
+{
+    public class Configuration
+    {
+        public string Host { get; set; } = "127.0.0.1";
+
+        public int Port { get; set; } = 48969;
+    }
+}